Rename inject-context-app files to react-context-app.
